Add option to process and send the details template as JSON
The details template option is super useful in reducing the clutter sent
through to pagerduty (the full event json is very verbose), but it gets sent as
text and so all the structure is lost on the Pagerduty side which means you
lose the ability to use custom details in event rules, etc.

This commit adds a optional `--send-details-as-json` flag which will try to
turn the processed details template string into a struct/json.

There isn't really any error handling though... if it fails to parse as
json you'll just get no custom details on the pagerduty side. That could
be improved (somehow?) to default back to sending text.

The rest of this commit message is more "notes to self".

- Since the `details` variable in `getDetails` is an `interface{}` it can be
anything
- So when using a template it returns a `string`
- Or if not using a template it returns `*v2.Event` which is a struct
- So we need to parse the returned templated string as a struct (unless we dive
much deeper and figure out doing this within templating itself)

This means we can use a template like this:

	--send-details-as-json --details-template "{\"check\": {\"name\": \"{{.Check.Name}}\", \"status\": {{.Check.Status}}, \"issued\": {{.Check.Issued}}, \"executed\": {{.Check.Executed}}, \"last_ok\": {{.Check.LastOK}}}, \"entity\": {\"name\": \"{{.Entity.Name}}\"}}" < event.json

Where it gets a bit trickier is trying to use labels/annotations. I.e. this
DOES NOT work:

	--send-details-as-json --details-template "{\"check\": {\"name\": \"{{.Check.Name}}\", \"status\": {{.Check.Status}}, \"issued\": {{.Check.Issued}}, \"executed\": {{.Check.Executed}}, \"last_ok\": {{.Check.LastOK}}, \"labels\": {{.Check.Labels}}}, \"entity\": {\"name\": \"{{.Entity.Name}}\", \"annotations\": {{.Entity.Annotations}}, \"subscriptions\": \"{{.Entity.Subscriptions}}\"}}" < event.json

Neither, necessarily, does quoting the labels and annotation fields
since within those there could be quotes anyway.

HOWEVER, although not documented, we can extract specific fields from
labels/annotations, which is good enough for building up a template.
I.e. this (simpler example) works:

	--send-details-as-json --details-template "{\"deployment_id\": \"{{.Check.Labels.deployment_id}}\"}" < event.json

Signed-off-by: atomicules <base@atomicules.co.uk>